==========Map Layer ============================ Added a speedvariation map, defined by colours. Green => high speed
===========Loader =============================
Loading screen shows up when the colours is being computed.
The map doesn't load before a while after though. This may be due to
googles own api, needing time to calculate after it gets data.

==========Input Fields===========================
Also added fields that enables the user to change the threshold for colours, as well as the opacity of the lines. Can be done with sliders.